fix(snack-runtime): pull `ExpoRouterEntry` out of `snack-runtime`
